infra: extract the L3/L4 tuple hash into a helper
The bond output node computes an RSS compatible hash from the L3/L4
tuple in order to pick a member. The same hash is needed elsewhere for
packets that carry no RSS hash at all, so move the tuple extraction to
flow_hash_l3l4() and the softrss call to flow_hash_words().

No functional change.
