+++ modules/infra/datapath/bond_output.c
@@ -2,17 +2,13 @@
 // Copyright (c) 2025 Robin Jarry
 
 #include "bond.h"
+#include "flow_hash.h"
 #include "graph.h"
 #include "iface.h"
 #include "mbuf.h"
 #include "rxtx.h"
 
 #include <rte_ether.h>
-#include <rte_ip4.h>
-#include <rte_ip6.h>
-#include <rte_tcp.h>
-#include <rte_thash.h>
-#include <rte_udp.h>
 
 #include <stdint.h>
 
@@ -34,28 +30,13 @@ static int bond_trace_format(char *buf, size_t len, const void *data, size_t /*d
 
 static inline const struct iface *
 hash_tx_member(const struct rte_mbuf *m, const struct iface_info_bond *bond) {
-	static const uint8_t rss_key[] = {
-		0x6d, 0x5a, 0x56, 0xda, 0x25, 0x5b, 0x0e, 0xc2, 0x41, 0x67, 0x25, 0x3d, 0x43, 0xa3,
-		0x8f, 0xb0, 0xd0, 0xca, 0x2b, 0xcb, 0xae, 0x7b, 0x30, 0xb4, 0x77, 0xcb, 0x2d, 0xa3,
-		0x80, 0x30, 0xf2, 0x0c, 0x6a, 0x42, 0xb7, 0x3b, 0xbe, 0xac, 0x01, 0xfa,
-	};
 	union {
 		uint32_t u32;
 		struct {
 			struct rte_ether_addr mac;
 			rte_be16_t vlan_id;
 		} l2;
-		struct rte_ipv4_tuple v4;
-		struct rte_ipv6_tuple v6;
 	} tuple;
-	union {
-		const struct rte_ipv4_hdr *ip4;
-		const struct rte_ipv6_hdr *ip6;
-	} l3;
-	union {
-		const struct rte_udp_hdr *udp;
-		const struct rte_tcp_hdr *tcp;
-	} l4;
 	const struct rte_ether_hdr *eth;
 	const struct rte_vlan_hdr *vlan;
 	uint32_t l3_offset, len, hash;
@@ -96,85 +77,16 @@ hash_tx_member(const struct rte_mbuf *m, const struct iface_info_bond *bond) {
 			eth_type = eth->ether_type;
 			l3_offset = sizeof(*eth);
 		}
-		switch (eth_type) {
-		case RTE_BE16(RTE_ETHER_TYPE_IPV4): {
-			l3.ip4 = rte_pktmbuf_mtod_offset(m, struct rte_ipv4_hdr *, l3_offset);
-			tuple.v4.src_addr = l3.ip4->src_addr;
-			tuple.v4.dst_addr = l3.ip4->dst_addr;
-			switch (l3.ip4->next_proto_id) {
-			case IPPROTO_UDP:
-				if (l3.ip4->fragment_offset == 0) {
-					l4.udp = rte_pktmbuf_mtod_offset(
-						m,
-						struct rte_udp_hdr *,
-						l3_offset + rte_ipv4_hdr_len(l3.ip4)
-					);
-					tuple.v4.sport = l4.udp->src_port;
-					tuple.v4.dport = l4.udp->dst_port;
-				} else {
-					// ignore UDP header for IP fragments
-					tuple.v4.sport = 0;
-					tuple.v4.dport = 0;
-				}
-				break;
-			case IPPROTO_TCP:
-				if (l3.ip4->fragment_offset == 0) {
-					l4.tcp = rte_pktmbuf_mtod_offset(
-						m,
-						struct rte_tcp_hdr *,
-						l3_offset + rte_ipv4_hdr_len(l3.ip4)
-					);
-					tuple.v4.sport = l4.tcp->src_port;
-					tuple.v4.dport = l4.tcp->dst_port;
-				} else {
-					// ignore TCP header for IP fragments
-					tuple.v4.sport = 0;
-					tuple.v4.dport = 0;
-				}
-				break;
-			default:
-				tuple.v4.sport = 0;
-				tuple.v4.dport = 0;
-			}
-			len = sizeof(tuple.v4);
-			break;
-		}
-		case RTE_BE16(RTE_ETHER_TYPE_IPV6): {
-			l3.ip6 = rte_pktmbuf_mtod_offset(m, const struct rte_ipv6_hdr *, l3_offset);
-			tuple.v6.src_addr = l3.ip6->src_addr;
-			tuple.v6.dst_addr = l3.ip6->dst_addr;
-			switch (l3.ip6->proto) {
-			case IPPROTO_UDP:
-				l4.udp = rte_pktmbuf_mtod_offset(
-					m, struct rte_udp_hdr *, l3_offset + sizeof(*l3.ip6)
-				);
-				tuple.v6.sport = l4.udp->src_port;
-				tuple.v6.dport = l4.udp->dst_port;
-				break;
-			case IPPROTO_TCP:
-				l4.tcp = rte_pktmbuf_mtod_offset(
-					m, struct rte_tcp_hdr *, l3_offset + sizeof(*l3.ip6)
-				);
-				tuple.v6.sport = l4.tcp->src_port;
-				tuple.v6.dport = l4.tcp->dst_port;
-				break;
-			default:
-				tuple.v6.sport = 0;
-				tuple.v6.dport = 0;
-			}
-			len = sizeof(tuple.v6);
-			break;
-		}
-		default:
-			len = sizeof(tuple.l2);
-			break;
-		}
+		if (flow_hash_l3l4(m, l3_offset, eth_type, &hash))
+			goto out;
+		// not an IP packet, fall back on the L2 tuple
+		len = sizeof(tuple.l2);
 		break;
 	default:
 		return NULL;
 	}
 
-	hash = rte_softrss_be(&tuple.u32, len / sizeof(uint32_t), rss_key);
+	hash = flow_hash_words(&tuple.u32, len / sizeof(uint32_t));
 out:
 	member = bond->redirection_table[hash % ARRAY_DIM(bond->redirection_table)];
 	if (member < bond->n_members)
